tpcdump: Update to 4.99.4
Prior to this MFC, the stable/13 tcpdump was at version 4.9.3, which was
released over five years ago.  Several CVEs have been identified in that
version.  For a detailed list, see
https://www.tcpdump.org/public-cve-list.txt.

Reviewed by:	kp (print-pf changes)
Sponsored by:	The FreeBSD Foundation

(cherry picked from commit 1cdec2832d051ddcb1417f3f2601e6212aff2613)
(cherry picked from commit ee67461e56828dd1f8de165947ba83f6d9148a87)
(cherry picked from commit a75d93a74d55d371876f5444075de4ecd685297f)
(cherry picked from commit 171a7bbfc04885150401ab64d96793373a8b2061)
(cherry picked from commit 1ad8d2ee1f7dec1d747ec955a68fbbb362958315)
(cherry picked from commit 7e0a7ef95fac1183854cab662bd9afa4647422d6)
(cherry picked from commit 85247ee6a2ba1c2dd0053e9be9055efa4be1438e)

Commit 1cdec2832d051ddcb1417f3f2601e6212aff2613 has been adapted for
stable/13.  Changes to usr.sbin/ndp/ndp_netlink.c have been omitted
because 91fbe0819bb9 (ndp: convert ndp(8) to netlink) has not been
MFCed.

Commit ee67461e56828dd1f8de165947ba83f6d9148a87 has been adapted for
stable/13.  Most modifications were required because
4bf98559d9d6fa7c3571d26ed6f2b18823e3a30b was not MFCed to stable/13.
Also incorporate a small change from
4848eb3af2a91b133c4b70cb9b71dd92ffec7f46 to fix the build.

+++ contrib/tcpdump/addrtostr.c
@@ -37,10 +37,10 @@
  */
 
 #ifdef HAVE_CONFIG_H
-#include "config.h"
+#include <config.h>
 #endif
 
-#include <netdissect-stdinc.h>
+#include "netdissect-stdinc.h"
 #include "addrtostr.h"
 
 #include <stdio.h>
@@ -71,7 +71,7 @@ addrtostr (const void *src, char *dst, size_t size)
 	return NULL;
     }
     for (i = 0; i < 4; ++i) {
-    	int n = *srcaddr++;
+	int n = *srcaddr++;
 	int non_zerop = 0;
 
 	if (non_zerop || n / 100 > 0) {
@@ -127,7 +127,7 @@ addrtostr6 (const void *src, char *dst, size_t size)
   best.base = -1;
   cur.len = 0;
   cur.base  = -1;
-  for (i = 0; i < (int)(IN6ADDRSZ / INT16SZ); i++)
+  for (i = 0; i < (IN6ADDRSZ / INT16SZ); i++)
   {
     if (words[i] == 0)
     {
@@ -160,14 +160,14 @@ addrtostr6 (const void *src, char *dst, size_t size)
         *dp++ = c; \
         space_left--; \
     }
-  for (i = 0; i < (int)(IN6ADDRSZ / INT16SZ); i++)
+  for (i = 0; i < (IN6ADDRSZ / INT16SZ); i++)
   {
     /* Are we inside the best run of 0x00's?
      */
     if (best.base != -1 && i >= best.base && i < (best.base + best.len))
     {
       if (i == best.base)
-      	 APPEND_CHAR(':');
+	  APPEND_CHAR(':');
       continue;
     }
